Implement cve-id middleware
