+++ src/controller/cve-id.controller/index.js
@@ -3,6 +3,7 @@ const router = express.Router()
 const mw = require('../../middleware/middleware')
 const controller = require('./cve-id.controller')
 const { check, validationResult } = require('express-validator')
+const { parseGetParams, parsePostParams } = require('./cve-id.middleware')
 
 const CHOICES = ['REJECT', 'PUBLIC', 'RESERVED']
 
@@ -72,63 +73,6 @@ function parseError(req, res, next) {
     next();
 }
 
-function parseGetParams(req, res, next) {
-    if(!('query' in req.ctx)) {
-        req.ctx.query = {}
-    }
-    if(!('params' in req.ctx)) {
-        req.ctx.params = {}
-    }
-    if('id' in req.params) {
-        req.ctx.params.id = req.params.id
-    }
-    if('state' in req.query) {
-        req.ctx.query.state = req.query.state
-    }
-    if('cve_id_year' in req.query) {
-        req.ctx.query.cve_id_year = req.query.cve_id_year
-    }
-    if('time_reserved.lt' in req.query) {
-        req.ctx.query['time_reserved.lt'] = req.query['time_reserved.lt']
-    }
-    if('time_reserved.gt' in req.query) {
-        req.ctx.query['time_reserved.gt'] = req.query['time_reserved.gt']
-    }
-    next();
+module.exports = {
+    router
 }
-
-function parsePostParams(req, res, next) {
-    if(!('query' in req.ctx)) {
-        req.ctx.query = {}
-    }
-    if(!('params' in req.ctx)) {
-        req.ctx.params = {}
-    }
-    if('state' in req.query) {
-        req.ctx.query.state = req.query.state
-    }
-    if('id' in req.params) {
-        req.ctx.params.id = req.params.id
-    }
-    if('year' in req.params) {
-        req.ctx.params.year = req.params.year
-    }
-    if('amount' in req.query) {
-        req.ctx.query.amount = req.query.amount
-    }
-    if('batch_type' in req.query) {
-        req.ctx.query.batch_type = req.query.batch_type
-    }
-    if('short_name' in req.query) {
-        req.ctx.query.short_name = req.query.short_name
-    }
-    if('cve_year' in req.query) {
-        req.ctx.query.cve_year = req.query.cve_year
-    }
-    if('cna' in req.query) {
-        req.ctx.query.cna = req.query.cna
-    }
-    next();
-}
-
-module.exports = router
