+++ test-utils/index.js
@@ -5,6 +5,7 @@ const app = express()
 const orgController = require('../src/controller/org.controller/org.controller')
 const cveController = require('../src/controller/cve.controller/cve.controller')
 const cveIdController = require('../src/controller/cve-id.controller/cve-id.controller')
+const cveIdParams = require('../src/controller/cve-id.controller/cve-id.middleware')
 const cveIdControllerA = require('./reserveCaseA.non-sequential')
 const cveIdControllerB = require('./reserveCaseB.non-sequential')
 const cveIdControllerE = require('./reserveCaseE.non-sequential')
@@ -98,43 +99,43 @@ app.route('/api/test/cna/:shortname/user/:username/reset_secret')
 
 // get a cve id specs from a cve document
 app.route('/api/test/cve-id/:id')
-  .get(cveIdController.CVEID_GET_SINGLE)
+  .get(cveIdParams.parseGetParams, cveIdController.CVEID_GET_SINGLE)
 
 // get a filtered cve document
 app.route('/api/test/cve-id')
-  .get(cveIdController.CVEID_GET_FILTER)
+  .get(cveIdParams.parseGetParams, cveIdController.CVEID_GET_FILTER)
 
 // modify a cve (cna reassigning or cve id STATE change)
 app.route('/api/test/cve-id/:id')
-  .post(cveIdController.CVEID_UPDATE_SINGLE)
+  .post(cveIdParams.parsePostParams, cveIdController.CVEID_UPDATE_SINGLE)
 
 // reserve cve id(s) Base Case, Case C and Case D
 app.route('/api/test/cve-id')
-  .post(cveIdController.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdController.CVEID_RESERVE)
 
 // reserve cve id(s) Case A
 app.route('/api/test/cve-id/case/A')
-  .post(cveIdControllerA.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdControllerA.CVEID_RESERVE)
 
 // reserve cve id(s) Case B
 app.route('/api/test/cve-id/case/B')
-  .post(cveIdControllerB.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdControllerB.CVEID_RESERVE)
 
 // reserve cve id(s) Case E
 app.route('/api/test/cve-id/case/E')
-  .post(cveIdControllerE.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdControllerE.CVEID_RESERVE)
 
 // reserve cve id(s) Case F
 app.route('/api/test/cve-id/case/F')
-  .post(cveIdControllerF.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdControllerF.CVEID_RESERVE)
 
 // reserve cve id(s) Case G
 app.route('/api/test/cve-id/case/G')
-  .post(cveIdControllerG.CVEID_RESERVE)
+  .post(cveIdParams.parsePostParams, cveIdControllerG.CVEID_RESERVE)
 
 // create a cve id range document
 app.route('/api/test/cve-id-range/:year')
-  .post(cveIdController.CVEID_RANGE_CREATE)
+  .post(cveIdParams.parsePostParams, cveIdController.CVEID_RANGE_CREATE)
 
 //* ********API Routes Setup END*********** */
